requirements.txt
Biopython
pyvcf
mongoengine

## Testing
detox
tox
pytest
pytest-cov